backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.2
python-dotenv==1.0.0
bcrypt==4.0.1
werkzeug==2.3.7
mysql-connector-python==8.1.0

# Testing dependencies
pytest==7.4.3
pytest-flask==1.3.0
selenium==4.15.2
webdriver-manager==4.0.1
coverage==7.3.2
pytest-cov==4.1.0